// libs/shared-models/src/lib/enums/license-tier.enum.ts
export enum LicenseTier {
  FREE = 'FREE',
  PATREON = 'PATREON',
  PATREON_SUPPORTER = 'PATREON_SUPPORTER',
  PATREON_PLUS = 'PATREON_PLUS',
  PATREON_PRO = 'PATREON_PRO',
  PATREON_ULTIMATE = 'PATREON_ULTIMATE',
  COMMERCIAL_STARTER = 'COMMERCIAL_STARTER',
  COMMERCIAL_PRO = 'COMMERCIAL_PRO',
  COMMERCIAL_ENTERPRISE = 'COMMERCIAL_ENTERPRISE',
}